Add layout for concept scheme
