+++ src/components/nestedList.js
@@ -1,0 +1,20 @@
+import PropTypes from "prop-types"
+import React from "react"
+
+const NestedList = ({items, current}) => (
+  <ul>
+    {items.map(item => (
+      <li key={item.id}>
+        <a
+          className={item.id === current ? 'current' : ''}
+          href={`${item.id.replace('http:/', '').replace('#', '')}.html`}
+        >
+          {item.prefLabel[0].value}
+        </a>
+        {item.narrower && <NestedList items={item.narrower} current={current} />}
+      </li>
+    ))}
+  </ul>
+)
+
+export default NestedList
