+++ src/templates/ConceptScheme.js
@@ -1,30 +1,42 @@
 import React from 'react'
-import { graphql } from 'gatsby'
+import { css, jsx } from '@emotion/core'
+import NestedList from '../components/nestedList'
 
-const ConceptScheme = ({pageContext, data}) => (
-  <div className="ConceptScheme">
-    <h1>ConceptScheme</h1>
-    <pre>
-      {JSON.stringify(pageContext, null, 2)}
-      {JSON.stringify(data, null, 2)}
-    </pre>
+const ConceptScheme = ({pageContext}) => {
+
+  return (
+  <div className="Concept">
+
+    <div className="layout"
+     css={css`
+      display: flex;
+      max-height: 100vh;
+      padding: 20px;
+
+      a.current {
+        color: tomato;
+        font-weight: bold;
+      }
+
+      nav {
+        overflow: auto;
+        border-right: 1px solid black;
+      }
+
+      .content {
+        padding: 0 20px;
+      }
+
+    `}>
+    <nav>
+      <NestedList items={JSON.parse(pageContext.node.tree).hasTopConcept} />
+    </nav>
+    <div className="content">
+      <h1>{pageContext.node.title}</h1>
+      <span>{pageContext.node.id}</span>
+    </div>
+    </div>
   </div>
-)
+)}
 
 export default ConceptScheme
-
-export const query = graphql`
-  query($hasTopConcept: [String]!) {
-    topConcepts: allConcept(filter: {id: {in: $hasTopConcept}}) {
-      edges {
-        node {
-          id,
-          prefLabel {
-            value
-            language
-          }
-        }
-      }
-    }
-  }
-`
